[FUEGO] Update a few mechanisms (LiD/dodecane_wang/drm19/gri30)

+++ Support/Fuego/Mechanism/Models/LiDryer/LiDryer.cpp
@@ -11,8 +11,8 @@ namespace thermo
     double activation_units[21], prefactor_units[21], phase_units[21];
     int is_PD[21], troe_len[21], sri_len[21], nTB[21], *TBid[21];
     double *TB[21];
-    std::vector<std::vector<int>> kiv(21); 
-    std::vector<std::vector<int>> nuv(21); 
+    std::vector<std::vector<double>> kiv(21); 
+    std::vector<std::vector<double>> nuv(21); 
 
     double fwd_A_DEF[21], fwd_beta_DEF[21], fwd_Ea_DEF[21];
     double low_A_DEF[21], low_beta_DEF[21], low_Ea_DEF[21];
